Fix -Wconversion warning (unsigned short int -> unsigned int)

+++ src/planet.h
@@ -12,8 +12,8 @@ struct Planet {
     unsigned short int x;            // x-coordinate of this planet's sector in the galaxy
     unsigned short int y;            // y-coordinate of the planet's sector in the galaxy
     Player* owner;                   // planet owner
-    unsigned short int res_per_turn; // resources generated at every turn
-    unsigned short int res_total;    // total resources available on the planet
+    unsigned int res_per_turn;       // resources generated at every turn
+    unsigned int res_total;          // total resources available on the planet
 
     void (*destroy)(Planet*);        // frees the planet
 };
